LPFG-941: Merge branch 'LPFG-940-Create-DB-Tables'. Changed Long ids to ints. Changed LocalDateTime fields to Instant

+++ src/main/java/uk/gov/cslearning/record/api/BookingController.java
@@ -22,7 +22,7 @@ public class BookingController {
     }
 
     @GetMapping(value = "/event/{eventId}/booking/{bookingId}", produces = "application/json")
-    public ResponseEntity<BookingDto> getBooking(@PathVariable long bookingId) {
+    public ResponseEntity<BookingDto> getBooking(@PathVariable int bookingId) {
         Optional<BookingDto> booking = bookingService.find(bookingId);
 
         return booking
